new_structure/mysql_migration/requirements.txt
Flask-SQLAlchemy>=2.5.1\nPyMySQL>=1.0.2\nSQLAlchemy>=1.4.0\ncryptography>=3.4.8\nmysql-connector-python>=8.0.33\n